Feature: 增加新功能并修复一些问题.
