+++ src/components/index.js
@@ -3,7 +3,7 @@ import Wbutton from './w-button_fe'
 
 import Wheader from './w-header_fe'
 
-import LuoDateTimePicker from '@/components/LuoDateTimePicker/LuoDateTimePicker'
+import LuoDateTimePicker from '@/components/LuoDateTimePicker/component/LuoDateTimePicker'
 
 // 加上你们自己的导入语句导入自己的组件
 const components = {
